% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semboottools-package.R
\docType{package}
\name{semboottools-package}
\alias{semboottools}
\alias{semboottools-package}
\title{semboottools: Bootstrapping Helpers For Structural Equation Modelling}
\description{
A collection of helper functions for forming bootstrapping confidence intervals and examining bootstrap estimates in structural equation modelling. Currently supports models fitted by the 'lavaan' package by Rosseel (2012) \doi{ 10.18637/jss.v048.i02}.
}
\author{
\strong{Maintainer}: Wendie Yang \email{temp@nil.com} (\href{https://orcid.org/0009-0000-8388-6481}{ORCID})

Authors:
\itemize{
  \item Shu Fai Cheung \email{shufai.cheung@gmail.com} (\href{https://orcid.org/0000-0002-9871-9448}{ORCID})
}

}
\keyword{internal}
